broke up parse.go file an places parsers with their type
